+++ changelog.md
@@ -5,6 +5,7 @@ Upcoming version
 ================
 ## New features
 - Help buttons on dialogs
+- Optionally reload data files automatically when file is changed
 - Simplified diffusion analysis report
 - Continuing support for Spanish localization (Mabel Galai)
 - Python package: results are returned in a dictionary